LayoutTests/inspector/elements/update-shadowdom-expected.txt

This test confirms that updating the shadow dom is reflected to the Inspector.

- <div id="container">
    - <div id="host">
        - #shadow-root
      </div>
  </div>
